feat: add create doctor profile, edit doctor profile, create patient profile, approve doctor, filter doctors

+++ src/modules/doctor/doctor.controller.js
@@ -1,4 +1,5 @@
 import asyncHandler from '../../common/middleware/async-handler.js';
+import { ApiError } from '../../common/utils/api/api-error.js';
 import { ApiResponse } from '../../common/utils/api/api-response.js';
 
 export default class DoctorController {
@@ -30,15 +31,49 @@ export default class DoctorController {
     });
 
     update = asyncHandler(async (req, res) => {
+        console.log(req.body);
+
+        const existingDoctor = await this.doctorService.getDoctorById(
+            req.params.doctorId,
+        );
+
+        const isOwner =
+            String(existingDoctor.userId._id) === String(req.auth.userId);
+
+        const isAdmin = req.auth.role === 'ADMIN';
+
+        if (!isOwner && !isAdmin) {
+            throw ApiError.forbidden(
+                'You are not allowed to update this doctor profile',
+            );
+        }
+
         const doctor = await this.doctorService.updateDoctor(
             req.params.doctorId,
             req.body,
         );
+
         return ApiResponse.ok(res, 'Doctor updated successfully', doctor);
     });
 
     remove = asyncHandler(async (req, res) => {
+        const doctor = await this.doctorService.getDoctorById(
+            req.params.doctorId,
+        );
+
+        const isOwner = String(doctor.userId._id) === String(req.auth.userId);
+
+        const isAdmin = req.auth.role === 'ADMIN';
+
+        if (!isOwner && !isAdmin) {
+            // console.log('match: ', doctor.userId._id, '-', req.auth.userId);
+            throw ApiError.forbidden(
+                'You are not allowed to delete this doctor profile',
+            );
+        }
+
         await this.doctorService.deleteDoctor(req.params.doctorId);
+
         return ApiResponse.ok(res, 'Doctor deleted successfully');
     });
